% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_setup.R
\name{run_protein_blast}
\alias{run_protein_blast}
\title{Run BLASTp to find protein function}
\usage{
run_protein_blast(protein_seq, db_path)
}
\arguments{
\item{protein_seq}{A string of amino acids}

\item{db_path}{Path to a BLAST database (e.g., Swiss-Prot)}
}
\description{
Run BLASTp to find protein function
}
